requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.10
openai>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0